STYLE : 상세페이지 탭 스타일 수정
