--- src/custom/jh/scrollTop.ts
@@ -1,4 +1,0 @@
-export const scrollTop = () => {
-  const top = document.getElementById('page-container');
-  if(top) {top.scrollTop = 0;}
-}
